advertisement
Use the list of vehicles to write an advertisement. So that it prints something like: "Amazing Joe's Garage, we service cars, motorbikes, caravans and bikes.". (Hint: use a for loop.)

What if you add one more vehicle to the list, can you have that added to the advertisement without changing the code for question 7?